requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.0.3
structlog==23.2.0
openai==1.3.0
groq==0.4.1
langdetect==1.0.9
pydub==0.25.1
SpeechRecognition==3.10.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.0
rich==13.7.0
